% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reg.r
\name{bglm_predict}
\alias{bglm_predict}
\title{predict function for bglm object}
\usage{
bglm_predict(mf = stop("mf: model frame must be supplied"),
  object = stop("object: bglm object must be supplied"),
  type = stop("type: either \\"link\\", \\"response\\", \\"terms\\""))
}
\arguments{
\item{mf}{model frame}

\item{object}{a bglm object}

\item{type}{one of c("link", "response", "terms")}
}
\description{
predict function for bglm object
}

