% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reg.r
\name{blm}
\alias{blm}
\title{Function to carry out linear regression on a data_frame data object}
\usage{
blm(formula = stop("formula: not supplied"),
  data = stop("data: data not supplied"), control = list(),
  weights = NULL, offset = NULL)
}
\arguments{
\item{formula}{formula that defines your regression model}

\item{data}{data_frame object containing data for linear regression}

\item{control}{list of parameters for control() function}

\item{weights}{weights for the model}

\item{offset}{offsets for the model}
}
\description{
Function to carry out linear regression on a data_frame data object
}

