% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CreateFactor}
\alias{CreateFactor}
\title{creates factor from numeric vector and character vector as levels}
\usage{
CreateFactor(x, levels)
}
\arguments{
\item{x}{numeric vector containing the numeric indices of the levels}

\item{levels}{character vector levels}
}
\description{
The \code{CreateFactor} function creates a factor from a numeric 
				vector and a character vector for levels
}

