% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_matrix.r
\name{data_matrix}
\alias{data_matrix}
\title{function to create a data_frame object}
\usage{
data_matrix(data = stop("data: matrix must be supplied"),
  chunkSize = stop("chunkSize must be specified, a good number is 50000"),
  path = stop("path must be specified"), nCores = parallel::detectCores(),
  ...)
}
\arguments{
\item{data}{object to be converted into a data_matrix object}

\item{chunkSize}{number of rows to be used in each chunk}

\item{path}{character to folder where the object will be created}

\item{nCores}{the number of cores to use defaults to parallel::detectCores()}

\item{...}{not used at the moment}
}
\description{
function to create a data_matrix object. The data_matrix object
             is an object that is held on disk. It is written to a folder
             \code{path} on disk where the data is written to in blocks or
             chunks. The data is written in binary format using a C++ function
             in purely numerical data.
}
\details{
Creates a data_matrix object
}

