% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reg.r
\name{.control}
\alias{.control}
\title{Function for creating control parameters for the GLM fit}
\usage{
.control(epsilon = 1e-08, maxit = 25, trace = TRUE)
}
\arguments{
\item{epsilon}{defaults to 1E-8}

\item{maxit}{defaults 25 maximum number of iterations}

\item{trace}{defaults to TRUE}
}
\description{
Function for creating control parameters for the GLM fit
}
