% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.r
\name{family_}
\alias{family_}
\title{family function}
\usage{
family_(distr, link)
}
\arguments{
\item{distr}{distr character one of "binomial", "poisson", "gaussian", "quasipoisson", 
"quasibinomial", "Gamma", "inverse.gaussian", "quasi"}

\item{link}{function character}
}
\description{
family function
}

