% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.r
\name{inverse.gaussian_}
\alias{inverse.gaussian_}
\title{inverse.gaussian family function}
\usage{
inverse.gaussian_(link = "1/mu^2")
}
\arguments{
\item{link}{function character}
}
\description{
inverse.gaussian family function
}

