% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame.r
\name{load_data_frame}
\alias{load_data_frame}
\title{function to load data_frame object}
\usage{
load_data_frame(path = stop("path: to data_frame folder must be supplied"))
}
\arguments{
\item{path}{character to folder containing object}
}
\description{
function to load data_frame object
}

