% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_matrix.r
\name{load_data_matrix}
\alias{load_data_matrix}
\title{function to load data_frame object}
\usage{
load_data_matrix(path = stop("path: to data_matrix folder must be supplied"))
}
\arguments{
\item{path}{character to folder containing object}
}
\description{
function to load data_frame object
}

