% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mySeq}
\alias{mySeq}
\title{mySeq function to sequence integers}
\usage{
mySeq(start, end)
}
\arguments{
\item{start}{integer from where sequence should start}

\item{end}{integer where sequence should end}
}
\description{
a function to create a sequence of integers
}

