% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame.r
\docType{data}
\name{plasma}
\alias{plasma}
\title{plasma data from the HSAUR package}
\format{a data.frame}
\source{
\href{https://cran.r-project.org/package=HSAUR}{HSAUR package}
}
\usage{
data(plasma)
}
\description{
Dataset from the HSAUR package
}
\details{
...
}
\examples{
data(plasma)
head(plasma)
}
\references{
HSAUR R package
(\href{https://cran.r-project.org/package=HSAUR}{HSAUR package})
}
\keyword{datasets}

