% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reg.r
\name{print.bglm}
\alias{print.bglm}
\title{print function for the bglm object}
\usage{
\method{print}{bglm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{bglm object to be displayed}

\item{digits}{number of significant digits to use}

\item{...}{not yet used}
}
\description{
print function for the bglm object
}

