% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reg.r
\name{print.blm}
\alias{print.blm}
\title{print function for the blm object}
\usage{
\method{print}{blm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{blm object to be displayed}

\item{digits}{number of significant digits to use}

\item{...}{not yet used}
}
\description{
print function for the blm object
}

