% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame.r
\name{print.data_frame}
\alias{print.data_frame}
\title{print function for a data_frame}
\usage{
\method{print}{data_frame}(x, ...)
}
\arguments{
\item{x}{data_frame object to print}

\item{...}{not used}
}
\description{
print function for a data_frame
}

