% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_matrix.r
\name{print.data_matrix}
\alias{print.data_matrix}
\title{print function for a data_matrix}
\usage{
\method{print}{data_matrix}(x, ...)
}
\arguments{
\item{x}{data_matrix object to print}

\item{...}{not used}
}
\description{
print function for a data_matrix
}

