% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reg.r
\name{print.summary.blm}
\alias{print.summary.blm}
\title{Function to print the summary object from the blm object}
\usage{
\method{print}{summary.blm}(x, digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{x}{summary blm object}

\item{digits}{- the digits to be displayed}

\item{signif.stars}{passed to printCoefmat}

\item{...}{arguments passed to \code{printCoefmat()} function}
}
\description{
Function to print the summary object from the blm object
}

