% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reg.r
\name{process_bglm_block}
\alias{process_bglm_block}
\title{Function to print the summary object from the blm object}
\usage{
process_bglm_block(mf, formula, mmCall, family, offset, weights, start, niter,
  etastart, mustart)
}
\arguments{
\item{mf}{the data block to be processed}

\item{formula}{the formula of for the model}

\item{mmCall}{the call object of the model}

\item{family}{the family object for the model}

\item{offset}{the model offset}

\item{weights}{the model weights}

\item{start}{the starting coefficient estimates}

\item{niter}{the current number of iterations}

\item{etastart}{the start for eta}

\item{mustart}{the start for mu}
}
\description{
Function to print the summary object from the blm object
}

