% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.r
\name{quasi_}
\alias{quasi_}
\title{quasi family function}
\usage{
quasi_(link = "identity", variance = "constant")
}
\arguments{
\item{link}{function character}

\item{variance}{choice character}
}
\description{
quasi family function
}

