% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.r
\name{quasibinomial_}
\alias{quasibinomial_}
\title{quasibinomial family function}
\usage{
quasibinomial_(link = "logit")
}
\arguments{
\item{link}{function character}
}
\description{
quasibinomial family function
}

