% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{r_bind}
\alias{r_bind}
\title{row binding for benchmarking ...}
\usage{
r_bind(x, y)
}
\arguments{
\item{x}{first matrix to be bound together}

\item{y}{second matrix to be bound together}
}
\description{
row binding for benchmarking
}

