% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_df_block}
\alias{read_df_block}
\title{read data frame block from file}
\usage{
read_df_block(size, filePath, df, ncol, factors, factor_indices)
}
\arguments{
\item{size}{number of elements in the block}

\item{filePath}{path to where the block is stored}

\item{df}{an empty list having the same number of elements as columns in the table}

\item{ncol}{number of columns in the dataframe block}

\item{factors}{list containing factors}

\item{factor_indices}{numeric vector containing the indicies that denote the factors}
}
\description{
read data frame block from file
}

