% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_df_blocks}
\alias{read_df_blocks}
\title{read multiple blocks of data frames from file}
\usage{
read_df_blocks(size, filePaths, df, ncols, factors, factor_indices)
}
\arguments{
\item{size}{number of elements in each block}

\item{filePaths}{path to where the blocks are stored}

\item{df}{an empty list having the same number of elements as columns in the table}

\item{ncols}{number of columns in the dataframe block}

\item{factors}{list containing factors}

\item{factor_indices}{numeric vector containing the indicies that denote the factors}
}
\description{
read multiple blocks of data frames from file
}

