% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_matrix_block}
\alias{read_matrix_block}
\title{read matrix block from file}
\usage{
read_matrix_block(filePath, size, ncol)
}
\arguments{
\item{filePath}{path to file where matrix should be read from}

\item{size}{total number of elements to be read}

\item{ncol}{number of columns in the matrix}
}
\description{
read matrix block from file
}

