% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_matrix_blocks}
\alias{read_matrix_blocks}
\title{read matrix blocks from file}
\usage{
read_matrix_blocks(filePaths, size, ncols)
}
\arguments{
\item{filePaths}{file paths from where the matrix blocks will be read}

\item{size}{numeric vector containing the number of elements in each block}

\item{ncols}{number of columns in the matrix}
}
\description{
read matrix blocks from file
}

