% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readNumericVector}
\alias{readNumericVector}
\title{reads numeric vector to file}
\usage{
readNumericVector(size, filePath)
}
\arguments{
\item{size}{the length of the numeric vector}

\item{filePath}{dependent variable}
}
\description{
reads numeric vector to file
}

