% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reg.r
\name{sum_bglm_block}
\alias{sum_bglm_block}
\title{The reduction function for the algorithm}
\usage{
sum_bglm_block(x1, x2)
}
\arguments{
\item{x1}{the first list object to be reduced}

\item{x2}{the second list object to be reduced}
}
\description{
The reduction function for the algorithm
}

