% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reg.r
\name{summary.bglm}
\alias{summary.bglm}
\title{summary function for the bglm object}
\usage{
\method{summary}{bglm}(object, ...)
}
\arguments{
\item{object}{bglm object to be summarized}

\item{...}{not used}
}
\description{
summary function for the bglm object
}

