% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_reg.r
\name{summary.blm}
\alias{summary.blm}
\title{summary function for the blm object}
\usage{
\method{summary}{blm}(object, ...)
}
\arguments{
\item{object}{blm object to be summarized}

\item{...}{not used}
}
\description{
summary function for the blm object
}

