% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SVD}
\alias{SVD}
\title{Singular value decomposition of the aggregated list from XWXMatrix(W) functions}
\usage{
SVD(out, epsilon)
}
\arguments{
\item{out}{list containing requisite computed values}

\item{epsilon}{either machine epsilon or user depermined epsilon}
}
\description{
Singular value decomposition of the aggregated list from XWXMatrix(W) functions
}

