% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{write_numeric_vector}
\alias{write_numeric_vector}
\title{writes numeric vector to file}
\usage{
write_numeric_vector(v, filePath)
}
\arguments{
\item{v}{numeric vector to be written to file}

\item{filePath}{path to file where the numeric vector should be written}
}
\description{
writes numeric vector to file
}

