% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{writeNumericVector}
\alias{writeNumericVector}
\title{writes numeric vector to file}
\usage{
writeNumericVector(v, filePath)
}
\arguments{
\item{v}{numeric vector}

\item{filePath}{dependent variable}
}
\description{
writes numeric vector to file
}

