% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{XWXMatrix}
\alias{XWXMatrix}
\title{Calculation of iterative regression components}
\usage{
XWXMatrix(X, y)
}
\arguments{
\item{X}{design matrix}

\item{y}{dependent variable}
}
\description{
Calculation of iterative regression components
}

