% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{XWXMatrixW}
\alias{XWXMatrixW}
\title{Calculation of iterative regression components}
\usage{
XWXMatrixW(X, y, W)
}
\arguments{
\item{X}{design matrix}

\item{y}{dependent variable}

\item{W}{weights}
}
\description{
Calculation of iterative regression components
}

